pyproject.toml
[project]
name = "sdk-exam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.8.0",
    "openai-agents>=0.3.0",
    "python-dotenv>=1.1.1",
]
